update library project
- update library assembly version
- update  IStepManagerand StepManger
- add x64 configuration for library and test project

+++ SimpleStepWriterSolution/SimpleStepWriter/IStepManager.cs
@@ -2,7 +2,7 @@
 {
     internal interface IStepManager
     {
-        long NextId { get; set; }
-        long ObjectIndex { get; set; }
+        int NextId { get; set; }        
+        int ObjectIndex { get; set; }
     }
 }
